# Use channel to push to policy