--- a/ring.go
+++ b/ring.go
@@ -49,6 +49,8 @@ func newRingStripe(config *ringConfig) *ringStripe {
 	}
 }
 
+var numPendingGo int32
+
 // Push appends an item in the ring buffer and drains (copies items and
 // sends to Consumer) if full.
 func (s *ringStripe) Push(item uint64) {
@@ -56,8 +58,13 @@ func (s *ringStripe) Push(item uint64) {
 	// if we should drain
 	if len(s.data) >= s.capacity {
 		// Send elements to consumer. Create a new one.
-		go s.consumer.Push(s.data)
-		s.data = make([]uint64, 0, s.capacity)
+		if pending := atomic.AddInt32(&numPendingGo, 1); pending <= 1 {
+			go s.consumer.Push(s.data)
+			s.data = make([]uint64, 0, s.capacity)
+		} else {
+			atomic.AddInt32(&numPendingGo, -1)
+			s.data = s.data[:0]
+		}
 	}
 }
 
@@ -126,6 +133,7 @@ func pushLossy(b *ringBuffer, item uint64) {
 	stripe := b.pool.Get().(*ringStripe)
 	stripe.Push(item)
 	b.pool.Put(stripe)
+	atomic.AddInt32(&numPendingGo, -1)
 }
 
 func pushLossless(b *ringBuffer, item uint64) {